speech-processing/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai-whisper==20231117
numpy==1.25.0
torch==2.3.1
gunicorn==23.0.0
pytest
httpx